+++ package.json
@@ -14,7 +14,8 @@
   },
   "dependencies": {
     "@expo/vector-icons": "^14.0.2",
-    "@react-navigation/native": "^7.0.14",
+    "@react-navigation/native": "^7.1.6",
+    "@react-navigation/native-stack": "^7.3.10",
     "expo": "~52.0.43",
     "expo-camera": "~16.0.18",
     "expo-font": "~13.0.4",
@@ -31,7 +32,7 @@
     "react-native": "0.76.9",
     "react-native-loading-spinner-overlay": "^3.0.1",
     "react-native-reanimated": "~3.16.1",
-    "react-native-safe-area-context": "4.12.0",
+    "react-native-safe-area-context": "^4.12.0",
     "react-native-screens": "~4.4.0",
     "react-native-vision-camera": "^4.6.4",
     "react-native-web": "~0.19.13"
